+++ docs/script/search_index.js
@@ -612,6 +612,12 @@ window.esdocSearchIndex = [
     "class"
   ],
   [
+    "formiojs/src/utils/utils.js~addtemplatehash",
+    "function/index.html#static-function-addTemplateHash",
+    "<span>addTemplateHash</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
+    "function"
+  ],
+  [
     "formiojs/src/providers/storage/base64.js~base64",
     "function/index.html#static-function-base64",
     "<span>base64</span> <span class=\"search-result-import-path\">formiojs/src/providers/storage/base64.js</span>",
@@ -1927,13 +1933,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/formio.js~formio.token",
-    "class/src/Formio.js~Formio.html#static-get-token",
+    "class/src/Formio.js~Formio.html#static-set-token",
     "src/Formio.js~Formio.token",
     "member"
   ],
   [
     "src/formio.js~formio.token",
-    "class/src/Formio.js~Formio.html#static-set-token",
+    "class/src/Formio.js~Formio.html#static-get-token",
     "src/Formio.js~Formio.token",
     "member"
   ],
@@ -2389,13 +2395,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/webform.js~webform#loading",
-    "class/src/Webform.js~Webform.html#instance-get-loading",
+    "class/src/Webform.js~Webform.html#instance-set-loading",
     "src/Webform.js~Webform#loading",
     "member"
   ],
   [
     "src/webform.js~webform#loading",
-    "class/src/Webform.js~Webform.html#instance-set-loading",
+    "class/src/Webform.js~Webform.html#instance-get-loading",
     "src/Webform.js~Webform#loading",
     "member"
   ],
@@ -2563,13 +2569,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/webform.js~webform#src",
-    "class/src/Webform.js~Webform.html#instance-get-src",
+    "class/src/Webform.js~Webform.html#instance-set-src",
     "src/Webform.js~Webform#src",
     "member"
   ],
   [
     "src/webform.js~webform#src",
-    "class/src/Webform.js~Webform.html#instance-set-src",
+    "class/src/Webform.js~Webform.html#instance-get-src",
     "src/Webform.js~Webform#src",
     "member"
   ],
@@ -4189,13 +4195,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/components/base/base.js~basecomponent#label",
-    "class/src/components/base/Base.js~BaseComponent.html#instance-set-label",
+    "class/src/components/base/Base.js~BaseComponent.html#instance-get-label",
     "src/components/base/Base.js~BaseComponent#label",
     "member"
   ],
   [
     "src/components/base/base.js~basecomponent#label",
-    "class/src/components/base/Base.js~BaseComponent.html#instance-get-label",
+    "class/src/components/base/Base.js~BaseComponent.html#instance-set-label",
     "src/components/base/Base.js~BaseComponent#label",
     "member"
   ],
@@ -4939,13 +4945,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/components/checkbox/checkbox.js~checkboxcomponent#datavalue",
-    "class/src/components/checkbox/Checkbox.js~CheckBoxComponent.html#instance-get-dataValue",
+    "class/src/components/checkbox/Checkbox.js~CheckBoxComponent.html#instance-set-dataValue",
     "src/components/checkbox/Checkbox.js~CheckBoxComponent#dataValue",
     "member"
   ],
   [
     "src/components/checkbox/checkbox.js~checkboxcomponent#datavalue",
-    "class/src/components/checkbox/Checkbox.js~CheckBoxComponent.html#instance-set-dataValue",
+    "class/src/components/checkbox/Checkbox.js~CheckBoxComponent.html#instance-get-dataValue",
     "src/components/checkbox/Checkbox.js~CheckBoxComponent#dataValue",
     "member"
   ],
@@ -5616,6 +5622,18 @@ window.esdocSearchIndex = [
     "method"
   ],
   [
+    "src/components/datagrid/datagrid.js~datagridcomponent#buildgroup",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-buildGroup",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#buildGroup",
+    "method"
+  ],
+  [
+    "src/components/datagrid/datagrid.js~datagridcomponent#buildgroups",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-buildGroups",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#buildGroups",
+    "method"
+  ],
+  [
     "src/components/datagrid/datagrid.js~datagridcomponent#buildrow",
     "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-buildRow",
     "src/components/datagrid/DataGrid.js~DataGridComponent#buildRow",
@@ -5700,6 +5718,12 @@ window.esdocSearchIndex = [
     "method"
   ],
   [
+    "src/components/datagrid/datagrid.js~datagridcomponent#getrowchunks",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-getRowChunks",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#getRowChunks",
+    "method"
+  ],
+  [
     "src/components/datagrid/datagrid.js~datagridcomponent#getvalue",
     "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-getValue",
     "src/components/datagrid/DataGrid.js~DataGridComponent#getValue",
@@ -5736,6 +5760,12 @@ window.esdocSearchIndex = [
     "method"
   ],
   [
+    "src/components/datagrid/datagrid.js~datagridcomponent#hasrowgroups",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-hasRowGroups",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#hasRowGroups",
+    "method"
+  ],
+  [
     "src/components/datagrid/datagrid.js~datagridcomponent#hastopsubmit",
     "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-hasTopSubmit",
     "src/components/datagrid/DataGrid.js~DataGridComponent#hasTopSubmit",
@@ -5754,6 +5784,18 @@ window.esdocSearchIndex = [
     "member"
   ],
   [
+    "src/components/datagrid/datagrid.js~datagridcomponent#removebutton",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-removeButton",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#removeButton",
+    "method"
+  ],
+  [
+    "src/components/datagrid/datagrid.js~datagridcomponent#removebuttonsmall",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-removeButtonSmall",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#removeButtonSmall",
+    "method"
+  ],
+  [
     "src/components/datagrid/datagrid.js~datagridcomponent#restorecomponentscontext",
     "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-restoreComponentsContext",
     "src/components/datagrid/DataGrid.js~DataGridComponent#restoreComponentsContext",
@@ -5764,6 +5806,12 @@ window.esdocSearchIndex = [
     "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-member-rows",
     "src/components/datagrid/DataGrid.js~DataGridComponent#rows",
     "member"
+  ],
+  [
+    "src/components/datagrid/datagrid.js~datagridcomponent#setstaticvalue",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-setStaticValue",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#setStaticValue",
+    "method"
   ],
   [
     "src/components/datagrid/datagrid.js~datagridcomponent#setvalue",
@@ -5782,6 +5830,12 @@ window.esdocSearchIndex = [
     "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-member-tableElement",
     "src/components/datagrid/DataGrid.js~DataGridComponent#tableElement",
     "member"
+  ],
+  [
+    "src/components/datagrid/datagrid.js~datagridcomponent#totalrowsnumber",
+    "class/src/components/datagrid/DataGrid.js~DataGridComponent.html#instance-method-totalRowsNumber",
+    "src/components/datagrid/DataGrid.js~DataGridComponent#totalRowsNumber",
+    "method"
   ],
   [
     "src/components/datagrid/datagrid.js~datagridcomponent#type",
@@ -6648,6 +6702,12 @@ window.esdocSearchIndex = [
     "file"
   ],
   [
+    "src/components/editgrid/editform/editgrid.edit.data.js",
+    "file/src/components/editgrid/editForm/EditGrid.edit.data.js.html",
+    "src/components/editgrid/editForm/EditGrid.edit.data.js",
+    "file"
+  ],
+  [
     "src/components/editgrid/editform/editgrid.edit.templates.js",
     "file/src/components/editgrid/editForm/EditGrid.edit.templates.js.html",
     "src/components/editgrid/editForm/EditGrid.edit.templates.js",
@@ -7423,13 +7483,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/components/form/form.js~formcomponent#parentvisible",
-    "class/src/components/form/Form.js~FormComponent.html#instance-set-parentVisible",
+    "class/src/components/form/Form.js~FormComponent.html#instance-get-parentVisible",
     "src/components/form/Form.js~FormComponent#parentVisible",
     "member"
   ],
   [
     "src/components/form/form.js~formcomponent#parentvisible",
-    "class/src/components/form/Form.js~FormComponent.html#instance-get-parentVisible",
+    "class/src/components/form/Form.js~FormComponent.html#instance-set-parentVisible",
     "src/components/form/Form.js~FormComponent#parentVisible",
     "member"
   ],
@@ -8275,13 +8335,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/components/nested/nestedcomponent.js~nestedcomponent#visible",
-    "class/src/components/nested/NestedComponent.js~NestedComponent.html#instance-get-visible",
+    "class/src/components/nested/NestedComponent.js~NestedComponent.html#instance-set-visible",
     "src/components/nested/NestedComponent.js~NestedComponent#visible",
     "member"
   ],
   [
     "src/components/nested/nestedcomponent.js~nestedcomponent#visible",
-    "class/src/components/nested/NestedComponent.js~NestedComponent.html#instance-set-visible",
+    "class/src/components/nested/NestedComponent.js~NestedComponent.html#instance-get-visible",
     "src/components/nested/NestedComponent.js~NestedComponent#visible",
     "member"
   ],